Make timeout e2e test reliably exercise the watchdog
The most recent full-suite run failed the (now-tightened) timeout
assertion because the agent scope-reduced the vague refactor task and
finished in 50s instead of being killed by the 1-minute watchdog:

  Tool: finish({'success': True, 'explanation': 'Created utils.py with
         type hints and docstrings, and test_utils.py with 14 passing
         unit tests.'})

The test had been silently coasting on the lax "no comment found =
PASS" assertion. PR #624 tightened it to FAIL, which surfaced this:
the watchdog isn't actually being exercised because the task is too
soft and the timeout too long.

## Fix

Two changes, both needed:

- **Concrete uncompressible task**: replace "refactor every file to
  follow best practices..." with "create exactly 50 Python files
  named day_01.py..day_50.py with this exact body." The agent can
  no longer satisfy the requirement by writing one demonstrative
  file; it's all 50 or explicit failure.

- **Shorter watchdog**: cut timeout_minutes from 1 to 0.5. Even
  scope-reduced work doesn't fit in 30s. Required adding fractional
  support: `timeout_minutes` is now `float` in lib/config.py's
  ALLOWED_ARGS, and the bash arithmetic in the workflow's 5
  watchdog steps changes from `$((TIMEOUT_MINUTES*60))` (bash
  integer math, drops fractional values to 0) to
  `awk -v m="$TIMEOUT_MINUTES" 'BEGIN{print int(m*60)}'`
  (handles int and float, rounds down to whole seconds).

The float type also leaves the door open for users who want
timeout_minutes=1.5 or similar in production configs.

655 unit tests still pass.

+++ lib/config.py
@@ -89,7 +89,7 @@ DEFAULT_TIMEOUT_MINUTES = 120
 ALLOWED_ARGS = {
     "max_iterations": int,       # agent.max_iterations (code-writing stages)
     "max_design_iterations": int,  # delegate: budget for design/exploration stages
-    "timeout_minutes": int,      # agent.timeout_minutes
+    "timeout_minutes": float,    # agent.timeout_minutes (float so e2e tests can use sub-minute values)
     "extra_files": list,         # mode's extra_files
     "branch": str,               # agent.branch (target branch for PRs)
     "status_log_interval": int,       # rolling status log interval (0 = disabled)
